+++ scorecardpy/__init__.py
@@ -8,9 +8,10 @@ from scorecardpy.var_filter import var_filter
 from scorecardpy.woebin import (woebin, woebin_ply, woebin_plot, woebin_adj)
 from scorecardpy.perf import (perf_eva, perf_psi)
 from scorecardpy.scorecard import (scorecard, scorecard_ply)
+from scorecardpy.one_hot import one_hot
 
 
-__version__ = '0.1.7.1'
+__version__ = '0.1.7.2'
 
 __all__ = (
     germancredit,
@@ -19,5 +20,6 @@ __all__ = (
     var_filter,
     woebin, woebin_ply, woebin_plot, woebin_adj,
     perf_eva, perf_psi,
-    scorecard, scorecard_ply
+    scorecard, scorecard_ply,
+    one_hot
 )