+++ scorecardpy/woebin.py
@@ -720,15 +720,16 @@ def woebin(dt, y, x=None, breaks_list=None, special_values=None,
     '''
     WOE Binning
     ------
-    `woebin` generates optimal binning for numerical, factor and categorical variables 
-    using methods including tree-like segmentation or chi-square merge. 
-    woebin can also customizing breakpoints if the breaks_list or special_values was provided.
+    `woebin` generates optimal binning for numerical, factor and categorical 
+    variables using methods including tree-like segmentation or chi-square 
+    merge. woebin can also customizing breakpoints if the breaks_list or 
+    special_values was provided.
     
     The default woe is defined as ln(Distr_Bad_i/Distr_Good_i). If you 
     prefer ln(Distr_Good_i/Distr_Bad_i), please set the argument `positive` 
     as negative value, such as '0' or 'good'. If there is a zero frequency 
-    class when calcuating woe, the zero will replaced by 0.99/total_number 
-    to make the woe calculable.
+    class when calcuating woe, the zero will replaced by 0.99 to make the 
+    woe calculable.
     
     Params
     ------